Return a proper response object.
We need to be able to interrogate the response properly; in particular we
need access to the status code, and the GSS code wants to verify the
server's token.

Parse the body from JSON and return an object that is a bit more convenient
to work with than the Apache response object.