Handle expired tokens.
Convert a 401 response to a Bearer request into an error. Catch this
error, invalidate the offending token and retry the sequence.

I now have to resort to a slight hack to create the /token request: I
can only build a ResolvedRequest with an FPHttpRequest, and a one of
those requires a service UUID. For the /token request I don't have the
service UUID to hand, I've got the resolved service URL instead.